+++ zarrs/src/array/array_async_readable.rs
@@ -776,7 +776,7 @@ impl<TStorage: ?Sized + AsyncReadableStorageTraits + 'static> Array<TStorage> {
                 .clone()
                 .async_partial_decoder(input_handle, &chunk_representation, options)
                 .await?
-                .partial_decode_opt(&[chunk_subset.clone()], options)
+                .partial_decode(&[chunk_subset.clone()], options)
                 .await?
                 .remove(0)
                 .into_owned()